% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_lowercase.R
\name{to_lowercase}
\alias{to_lowercase}
\title{Change a uppercase string to an lowercase string}
\usage{
to_lowercase(string)
}
\arguments{
\item{string}{a string containing only uppercase letters}
}
\value{
the string written in lowercase
}
\description{
Change a uppercase string to an lowercase string
}
\examples{
to_uppercase("PONY")
}
