% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_uppercase.R
\name{to_uppercase}
\alias{to_uppercase}
\title{Change a lowercase string to an uppercase string}
\usage{
to_uppercase(string)
}
\arguments{
\item{string}{a string containing only lowercase letters}
}
\value{
the string written in uppercase
}
\description{
Change a lowercase string to an uppercase string
}
\examples{
to_uppercase("pony")
}
